Move LazyLogHandler to logger.handler package.
